Fix setting version in app settings
