+++ src/WebApp/Platform/Settings/AppSettings.cs
@@ -1,12 +1,11 @@
 ﻿using AirWeb.WebApp.Platform.PrintoutModels;
 using JetBrains.Annotations;
-using System.Reflection;
 
 namespace AirWeb.WebApp.Platform.Settings;
 
 internal static partial class AppSettings
 {
-    public static string Version { get; } = GetVersion();
+    public static string? Version { get; set; }
     public static Support SupportSettings { get; } = new();
     public static Raygun RaygunSettings { get; } = new();
     public static OrganizationInfo OrganizationInfo { get; } = new();
@@ -21,13 +20,5 @@ internal static partial class AppSettings
     public record Raygun
     {
         public string? ApiKey { get; [UsedImplicitly] init; }
-    }
-
-    private static string GetVersion()
-    {
-        var entryAssembly = Assembly.GetEntryAssembly();
-        var segments = (entryAssembly?.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?
-            .InformationalVersion ?? entryAssembly?.GetName().Version?.ToString() ?? "").Split('+');
-        return segments[0] + (segments.Length > 0 ? $"+{segments[1][..Math.Min(7, segments[1].Length)]}" : "");
     }
 }
